+++ reflective-client/src/types/RecentPost/RecentPost.tsx
@@ -1,0 +1,7 @@
+import { PostType } from "../types";
+
+export type PostCardProps = {
+  post: PostType;
+  index: number;
+  handlePost: (post_id: string, nickname: string, title: string) => void;
+};
